ref: implemented suggested chages
